Dodane testy jednostkowe, przeniesienie id w metodach patch z body do url, dodane validatory do niektórych schematów

+++ crud.py
@@ -2,6 +2,8 @@ import base64
 import imghdr
 import pathlib
 import datetime
+import os
+from django.http import Http404
 
 from sqlalchemy.orm import Session
 from fastapi import HTTPException
@@ -11,13 +13,19 @@ import models, schemas
 BASE_DIR = pathlib.Path(__file__).resolve().parent
 
 def get_user(db: Session, user_id: int):
-    return db.query(models.User).filter(models.User.id == user_id).first()
+    user = db.query(models.User).filter(models.User.id == user_id).first()
+    if user is None:
+        raise HTTPException(404, "User doesn't exist")
+    return user
 
 def check_user(db: Session, user_id: int):
     return db.query(models.User.id).filter(models.User.id == user_id).first() is not None
 
 def get_user_by_email(db: Session, email: str):
-    return db.query(models.User).filter(models.User.email == email).first()
+    user = db.query(models.User).filter(models.User.email == email).first()
+    if user is None:
+        raise HTTPException(404, "User doesn't exist")
+    return user
 
 def get_users(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.User).offset(skip).limit(limit).all()
@@ -30,25 +38,22 @@ def create_user(db: Session, user: schemas.UserCreate):
     db.refresh(db_user)
     return db_user
 
-def update_user(db: Session, user_id: int, user: schemas.User):
+def update_user(db: Session, user_id: int, user: schemas.UserEdit):
     db_user = get_user(db, user_id)
-    if db_user is not None:
-        for attr, value in user.dict(exclude_defaults=True).items():
-            setattr(db_user, attr, value)
-        db.commit()
-        return True
-    return False
+    for attr, value in user.dict(exclude_defaults=True).items():
+        setattr(db_user, attr, value)
+    db.commit()
 
 def delete_user(db: Session, user_id: int):
     db_user = get_user(db, user_id)
-    if db_user is not None:
-        db.delete(db_user)
-        db.commit()
-        return True
-    return False
+    db.delete(db_user)
+    db.commit()
 
 def get_car(db: Session, car_id: int):
-    return db.query(models.Car).filter_by(id=car_id).first()
+    car = db.query(models.Car).filter_by(id=car_id).first()
+    if car is None:
+        raise HTTPException(404, "Car doesn't exist!")
+    return car
 
 def check_car(db: Session, car_id: int):
     return db.query(models.Car.id).filter(models.Car.id == car_id).first() is not None
@@ -66,6 +71,11 @@ def save_img(path, file_bytes):
     with open(BASE_DIR / path, 'wb') as file_to_save:
         file_to_save.write(file_bytes)
 
+def delete_img(path):
+    try:
+        os.remove(BASE_DIR / path)
+    except:
+        pass
 
 def create_car(db: Session, car: schemas.CarCreate):
     # Obraz auta jest w base64, po odkodowaniu i zapisaniu pliku
@@ -84,46 +94,37 @@ def create_car(db: Session, car: schemas.CarCreate):
     db.commit()
     return db_car
 
-def update_car(db: Session, car : schemas.Car):
+def update_car(db: Session, car_id: int, car: schemas.CarEdit):
     if car.img is not None:
-        img_bytes, _ = decode_image(car.img)
-    db_car = get_car(db, car.id)
-    try:
-        if car.img is not None:
-            save_img(db_car.img, img_bytes)
-            car.img = None
-        for attr, value in car.dict(exclude_defaults=True).items():
-            setattr(db_car, attr, value)
-        #db_car.update(car.dict())
-        db.commit()
-    except:
-        raise HTTPException(404, "Car doesn't exist!")
+        img_bytes, ext = decode_image(car.img)
+    db_car = get_car(db, car_id)
+    img_path = f'media/cars/car_{db_car.id}.{ext}'
+    if car.img is not None:
+        delete_img(car.img)
+        #save_img(db_car.img, img_bytes)
+    save_img(img_path, img_bytes)
+    for attr, value in car.dict(exclude_defaults=True).items():
+        setattr(db_car, attr, value)
+    db_car.img = img_path
+    #db_car.update(car.dict())
+    db.commit()
 
 def delete_car(db: Session, car_id : int):
     db_car = get_car(db, car_id)
-    # if db_car is not None:
-    #     db.delete(db_car)
-    #     #db_car.delete()
-    #     db.commit()
-    #     return True
-    # return False
-    try:
-        db.delete(db_car)
-        db.commit()
-    except:
-        raise HTTPException(404, 'Car does not exist!')
+    db.delete(db_car)
+    db.commit()
         
 def get_active_rentals_by_car(db: Session, car_id: int):
     TODAY = datetime.date.today()
     rentals = db.query(models.Car).filter(models.Car.id == car_id).first().rentals
-    return rentals.filter(models.Rental.rental_end > TODAY).all()
+    return rentals.filter(models.Rental.rental_end >= TODAY).all()
 
 def get_active_rentals_by_user(db: Session, user_id: int):
     TODAY = datetime.date.today()
     rentals = db.query(models.User).filter(models.User.id == user_id).first().rentals
-    return rentals.filter(models.Rental.rental_end > TODAY).all()
+    return rentals.filter(models.Rental.rental_end >= TODAY).all()
 
-def get_finished_rentals_by_user(db: Session, user_id: int):
+def get_all_rentals_by_user(db: Session, user_id: int):
     return db.query(models.User).filter(models.User.id == user_id).first().rentals.all()
 
 def get_unpaid_rentals_by_user(db: Session, user_id: int):
@@ -131,32 +132,35 @@ def get_unpaid_rentals_by_user(db: Session, user_id: int):
     return rentals.filter(models.Rental.paid == False).all()
 
 def get_rental(db: Session, rental_id: int):
-    return db.query(models.Rental).filter(models.Rental.id == rental_id).first()
+    rental = db.query(models.Rental).filter(models.Rental.id == rental_id).first()
+    if rental is None:
+        raise HTTPException(404, "Rental doesn't exist")
+    return rental
 
 def rental_pay(db: Session, rental_id: int):
     rental = get_rental(db, rental_id)
-    try:
-        rental.paid = True
-        db.commit()
-    except:
-        raise HTTPException(404, "Rental does not exist!")
+    rental.paid = True
+    db.commit()
 
 def check_rental(db: Session, rental_id: int):
     return db.query(models.Rental.id).filter(models.Rental.id == rental_id).first() is not None
+
+def check_rental_active(db: Session, rental_id):
+    TODAY = datetime.date.today()
+    rental = db.query(models.Rental.id).filter(models.Rental.id == rental_id).first()
+    return rental.rental_end >= TODAY
 
 def create_rental(db: Session, user_id: int, rental: schemas.RentalCreate):
     db_rentals = get_active_rentals_by_car(db, rental.car_id)
     start = rental.rental_start
     end = rental.rental_end
 
-    DAY = datetime.timedelta(days=1)
+    #DAY = datetime.timedelta(days=1)
     for r in db_rentals: # Sprawdzanie czy okres wynajmu nie nachodzi na inny wynajem
-        if start >= r.rental_start and start <= (r.rental_end + DAY):
-            return 
-        elif end >= r.rental_start and end <= (r.rental_end + DAY):
-            return
-        elif start < r.rental_start and end > (r.rental_end + DAY):
-            return
+        if (start >= r.rental_start and start <= r.rental_end) or \
+        (end >= r.rental_start and end <= r.rental_end) or \
+        (start < r.rental_start and end > r.rental_end):
+            raise HTTPException(409, "Car is taken for given period!")
 
     db_item = models.Rental(**rental.dict(), user_id=user_id)
     db.add(db_item)
@@ -164,41 +168,40 @@ def create_rental(db: Session, user_id: int, rental: schemas.RentalCreate):
     db.refresh(db_item)
     return db_item
 
-def update_rental(db: Session, rental: schemas.RentalDateEdit):
-    current_rental = get_rental(db, rental.id)
-    try:
-        db_rentals = get_active_rentals_by_car(db, current_rental.car_id)
-        #db_rentals = db_rentals.filter(models.Rental.id != rental.id).all()
-        start = rental.rental_start
-        end = rental.rental_end
+def update_rental(db: Session, rental_id: int, rental: schemas.RentalBase):
+    current_rental = get_rental(db, rental_id)
+    if current_rental.paid:
+        raise HTTPException(409, "Rental has been paid for the whole period!")
+    db_rentals = get_active_rentals_by_car(db, current_rental.car_id)
+    #db_rentals = db_rentals.filter(models.Rental.id != rental.id).all()
+    start = rental.rental_start
+    end = rental.rental_end
 
-        DAY = datetime.timedelta(days=1)
-        for r in db_rentals: # Sprawdzanie czy okres wynajmu nie nachodzi na inny wynajem
-            if r.id == current_rental.id:
-                continue
-            if start >= r.rental_start and start <= (r.rental_end + DAY):
-                return 
-            elif end >= r.rental_start and end <= (r.rental_end + DAY):
-                return
-            elif start < r.rental_start and end > (r.rental_end + DAY):
-                return
-        for attr, value in rental.dict(exclude_defaults=True).items():
-            setattr(current_rental, attr, value)
-        db.commit()
-    except:
-        raise HTTPException(404, 'Rental does not exist!')
+    #DAY = datetime.timedelta(days=1)
+    for r in db_rentals: # Sprawdzanie czy okres wynajmu nie nachodzi na inny wynajem
+        if r.id == current_rental.id:
+            continue
+        if (start >= r.rental_start and start <= r.rental_end) or \
+        (end >= r.rental_start and end <= r.rental_end) or \
+        (start < r.rental_start and end > r.rental_end):
+            raise HTTPException(409, "Car is taken for given period!")
+
+    for attr, value in rental.dict(exclude_defaults=True).items():
+        setattr(current_rental, attr, value)
+    db.commit()
 
 def stop_rental(db: Session, rental_id: int):
     db_rental = get_rental(db, rental_id)
     TODAY = datetime.date.today()
     DAY = datetime.timedelta(days=1)
+    if db_rental.paid:
+        raise HTTPException(409, "Rental has been paid for the whole period!")
     if db_rental.rental_start > TODAY:
         db.delete(db_rental)
         db.commit()
-        return True
+        return
     db_rental.rental_end = TODAY + DAY
     db.commit()
-    return False
 
 # def delete_rental(db: Session, rental_id: int):
 #     db_rental = get_rental(db, rental_id)
